Data
This data is freely available from the NYC OpenData website (https://data.cityofnewyork.us/Transportation/Subway-Stations/arq3-7z49). It contains latitude and longitude of all subway stations in the five boroughs of NYC.

Simply click export, above the map, and choose .CSV. This data needs to be preprocessed, however. The latitude and longitude (the_geom) are reported in reversed order, and without a comma in between. To generate a usable data set, adjust the lat/lon to the proper order, and remove some extraneous variables that may not be needed. In order to make transitioning between the original Geocode program, and this modified Reverse-Geocode program, I have also renamed the location variable, "the_geom", to "Address."